+++ functions/level_1/four_bank_parser.py
