+++ tests/level_1/test_four_bank_parser.py
@@ -1,0 +1,5 @@
+from functions.level_1.four_bank_parser import BankCard, SmsMessage, Expense, parse_ineco_expense
+
+
+def test_parse_ineco_expense():
+    pass
