+++ tests/level_1/test_three_url_builder.py
@@ -1,0 +1,5 @@
+from functions.level_1.three_url_builder import build_url
+
+
+def test_build_url():
+    pass
