+++ tests/level_1/test_two_date_parser.py
@@ -1,0 +1,5 @@
+from functions.level_1.two_date_parser import compose_datetime_from
+
+
+def test_compose_datetime_from():
+    pass
