--- tests/test_five_title.py
@@ -1,5 +1,0 @@
-from functions.five_title import change_copy_item
-
-
-def test_change_copy_item():
-    pass
