--- tests/test_four_bank_parser.py
@@ -1,5 +1,0 @@
-from functions.four_bank_parser import BankCard, SmsMessage, Expense, parse_ineco_expense
-
-
-def test_parse_ineco_expense():
-    pass
