--- tests/test_one_gender.py
@@ -1,5 +1,0 @@
-from functions.one_gender import genderalize
-
-
-def test_genderalize():
-    pass
