--- tests/test_three_url_builder.py
@@ -1,5 +1,0 @@
-from functions.three_url_builder import build_url
-
-
-def test_build_url():
-    pass
