--- tests/test_two_date_parser.py
@@ -1,5 +1,0 @@
-from functions.two_date_parser import compose_datetime_from
-
-
-def test_compose_datetime_from():
-    pass
